Fix issue with chat_id in database

+++ schema.sql
@@ -1,21 +1,21 @@
 CREATE TABLE feedback (
-  user_id INTEGER NOT NULL,
+  user_id BIGINT NOT NULL,
   username VARCHAR,
   first_name VARCHAR,
   last_name VARCHAR,
-  chat_id INTEGER NOT NULL,
-  message_id INTEGER NOT NULL,
+  chat_id BIGINT NOT NULL,
+  message_id BIGINT NOT NULL,
   text TEXT,
   date TIMESTAMP
 );
 
 CREATE TABLE report (
-  user_id INTEGER NOT NULL,
+  user_id BIGINT NOT NULL,
   username VARCHAR,
   first_name VARCHAR,
   last_name VARCHAR,
-  chat_id INTEGER NOT NULL,
-  message_id INTEGER NOT NULL,
+  chat_id BIGINT NOT NULL,
+  message_id BIGINT NOT NULL,
   url VARCHAR,
   date TIMESTAMP
 );
